requirements.txt
astapi
uvicorn
requests
beautifulsoup4
lxml
pandas
python-multipart